requirements.txt
python-docx>=0.8.11
PyYAML>=6.0
langchain>=0.1.0
langchain-openai>=0.1.0
click>=8.0.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=1.0.0 